Swin_Tiny/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python
scikit-learn
matplotlib
seaborn
pandas
numpy
Pillow